pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reroute_backend"
version = "0.1.0"
description = "FastAPI backend for AI cycling coach"
authors = [{name = "Darragh Mahns"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.3.1",
  "httpx>=0.27.0",
  "structlog>=24.1.0",
  "python-jose[cryptography]>=3.3.0",
  "sqlalchemy>=2.0.29",
  "alembic>=1.13.1",
  "openai>=1.40.0",
  "pydantic-ai>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
  "respx>=0.20.2",
  "httpx>=0.27.0",
  "ruff>=0.4.2",
  "mypy>=1.9.0",
  "coverage[toml]>=7.4.4",
  "pytest-cov>=5.0.0"
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "W", "I", "N", "UP", "B"]

[tool.mypy]
python_version = "3.11"
plugins = []
strict = true
warn_unused_configs = true
